tpch and tpcc generator interface wip
