+++ src/benchmark/CMakeLists.txt
@@ -16,7 +16,7 @@ add_executable(
     operators/table_scan_benchmark.cpp
     operators/union_all_benchmark.cpp
     statistics/generate_table_statistics_benchmark.cpp
-    tpch_db_generator_benchmark.cpp
+    tpch_table_generator_benchmark.cpp
 )
 
 target_link_libraries(
@@ -36,7 +36,7 @@ add_executable(
 )
 target_link_libraries(
     hyriseBenchmarkPlayground
-    
+
     hyrise
     hyriseBenchmarkLib
     benchmark
