+++ src/benchmarklib/tpch/tpch_table_generator.cpp
@@ -1,4 +1,4 @@
-#include "tpch_db_generator.hpp"
+#include "tpch_table_generator.hpp"
 
 extern "C" {
 #include <dss.h>
@@ -8,6 +8,7 @@ extern "C" {
 
 #include <utility>
 
+#include "benchmark_utilities/table_builder.hpp"
 #include "storage/chunk.hpp"
 #include "storage/storage_manager.hpp"
 
@@ -42,88 +43,6 @@ const auto region_column_types = boost::hana::tuple<     int32_t,       std::str
 const auto region_column_names = boost::hana::make_tuple("r_regionkey", "r_name",    "r_comment");  // NOLINT
 
 // clang-format on
-
-/**
- * Helper to build a table with a static (specified by template args `ColumnTypes`) column type layout. Keeps a vector
- * for each column and appends values to them in append_row(). Automatically creates chunks in accordance with the
- * specified chunk size.
- *
- * No real need to tie this to TPCH, but atm it is only used here so that's where it resides.
- */
-template <typename... DataTypes>
-class TableBuilder {
- public:
-  template <typename... Strings>
-  TableBuilder(size_t chunk_size, const boost::hana::tuple<DataTypes...>& column_types,
-               const boost::hana::tuple<Strings...>& column_names, opossum::UseMvcc use_mvcc)
-      : _use_mvcc(use_mvcc) {
-    /**
-     * Create a tuple ((column_name0, column_type0), (column_name1, column_type1), ...) so we can iterate over the
-     * columns.
-     * fold_left as below does this in order, I think boost::hana::zip_with() doesn't, which is why I'm doing two steps
-     * here.
-     */
-    const auto column_names_and_data_types = boost::hana::zip_with(
-        [&](auto column_type, auto column_name) {
-          return boost::hana::make_tuple(column_name, opossum::data_type_from_type<decltype(column_type)>());
-        },
-        column_types, column_names);
-
-    // Iterate over the column types/names and create the columns.
-    opossum::TableColumnDefinitions column_definitions;
-    boost::hana::fold_left(column_names_and_data_types, column_definitions,
-                           [](auto& column_definitions, auto column_name_and_type) -> decltype(auto) {
-                             column_definitions.emplace_back(column_name_and_type[boost::hana::llong_c<0>],
-                                                             column_name_and_type[boost::hana::llong_c<1>]);
-                             return column_definitions;
-                           });
-    _table = std::make_shared<opossum::Table>(column_definitions, opossum::TableType::Data, chunk_size, use_mvcc);
-  }
-
-  std::shared_ptr<opossum::Table> finish_table() {
-    if (_current_chunk_row_count() > 0) {
-      _emit_chunk();
-    }
-
-    return _table;
-  }
-
-  void append_row(DataTypes&&... column_values) {
-    // Create a tuple ([&data_vector0, value0], ...)
-    auto vectors_and_values = boost::hana::zip_with(
-        [](auto& vector, auto&& value) { return boost::hana::make_tuple(std::reference_wrapper(vector), value); },
-        _data_vectors, boost::hana::make_tuple(std::forward<DataTypes>(column_values)...));
-
-    // Add the values to their respective data vector
-    boost::hana::for_each(vectors_and_values, [](auto vector_and_value) {
-      vector_and_value[boost::hana::llong_c<0>].get().push_back(vector_and_value[boost::hana::llong_c<1>]);
-    });
-
-    if (_current_chunk_row_count() >= _table->max_chunk_size()) {
-      _emit_chunk();
-    }
-  }
-
- private:
-  std::shared_ptr<opossum::Table> _table;
-  opossum::UseMvcc _use_mvcc;
-  boost::hana::tuple<opossum::pmr_concurrent_vector<DataTypes>...> _data_vectors;
-
-  size_t _current_chunk_row_count() const { return _data_vectors[boost::hana::llong_c<0>].size(); }
-
-  void _emit_chunk() {
-    opossum::Segments segments;
-
-    // Create a segment from each data vector and add it to the Chunk, then re-initialize the vector
-    boost::hana::for_each(_data_vectors, [&](auto&& vector) {
-      using T = typename std::decay_t<decltype(vector)>::value_type;
-      // reason for nolint: clang-tidy wants this to be a forward, but that doesn't work
-      segments.push_back(std::make_shared<opossum::ValueSegment<T>>(std::move(vector)));  // NOLINT
-      vector = std::decay_t<decltype(vector)>();
-    });
-    _table->append_chunk(segments);
-  }
-};
 
 std::unordered_map<opossum::TpchTable, std::underlying_type_t<opossum::TpchTable>> tpch_table_to_dbgen_id = {
     {opossum::TpchTable::Part, PART},     {opossum::TpchTable::PartSupp, PSUPP}, {opossum::TpchTable::Supplier, SUPP},
@@ -187,10 +106,10 @@ std::unordered_map<TpchTable, std::string> tpch_table_names = {
     {TpchTable::Customer, "customer"}, {TpchTable::Orders, "orders"},     {TpchTable::LineItem, "lineitem"},
     {TpchTable::Nation, "nation"},     {TpchTable::Region, "region"}};
 
-TpchDbGenerator::TpchDbGenerator(float scale_factor, uint32_t chunk_size)
-    : _scale_factor(scale_factor), _chunk_size(chunk_size) {}
+TpchTableGenerator::TpchTableGenerator(float scale_factor, uint32_t chunk_size, EncodingConfig config, bool store)
+    : AbstractBenchmarkTableGenerator(chunk_size, std::move(config), store), _scale_factor(scale_factor) {}
 
-std::unordered_map<TpchTable, std::shared_ptr<Table>> TpchDbGenerator::generate() {
+std::unordered_map<TpchTable, std::shared_ptr<Table>> TpchTableGenerator::generate() {
   TableBuilder customer_builder{_chunk_size, customer_column_types, customer_column_names, UseMvcc::Yes};
   TableBuilder order_builder{_chunk_size, order_column_types, order_column_names, UseMvcc::Yes};
   TableBuilder lineitem_builder{_chunk_size, lineitem_column_types, lineitem_column_names, UseMvcc::Yes};
@@ -297,12 +216,17 @@ std::unordered_map<TpchTable, std::shared_ptr<Table>> TpchDbGenerator::generate(
       {TpchTable::Nation, nation_builder.finish_table()},     {TpchTable::Region, region_builder.finish_table()}};
 }
 
-void TpchDbGenerator::generate_and_store() {
+void TpchTableGenerator::generate_and_store() {
   const auto tables = generate();
 
   for (auto& table : tables) {
     StorageManager::get().add_table(tpch_table_names.at(table.first), table.second);
   }
+}
+
+std::map<std::string, std::shared_ptr<opossum::Table>> TpchTableGenerator::_generate_all_tables() {
+  auto tables = generate();
+  return tables;
 }
 
 }  // namespace opossum
